// AdminApp/src/app/companyapi/models/tenant-user-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface TenantUserDto {
  adminPassword?: null | string;
  email?: null | string;
  firstName?: null | string;
  isActive?: boolean;
  isAdmin?: boolean;
  isUser?: boolean;
  lastName?: null | string;
  password?: null | string;
  tenant?: null | string;
  userId?: number;
  userName?: null | string;
}